who likes cows anyway
i kinda do but not really
